+++ app/src/main/java/me/shikhov/wlogtest/Application.java
@@ -1,0 +1,12 @@
+package me.shikhov.wlogtest;
+
+public class Application extends android.app.Application
+{
+    @Override
+    public void onCreate()
+    {
+        super.onCreate();
+
+        System.setProperty("wlog.logLevel", "info");
+    }
+}